Auto trajectory following works now
Can drive x,y and turn to angle, needs more tuning but it works pretty well. Turn to angle for driver control sort of works but likely needs to make a new controller each time the command is run because it looks like it briefly remembers to previous goal before turning to the new one.